requirements.txt
# Data Processing
numpy>=1.19.2
pandas>=1.2.0
pyarrow>=7.0.0

# Machine Learning
scikit-learn>=0.24.0
xgboost>=1.5.0
lightgbm>=3.3.0
pycaret==3.1.0

# MLOps
mlflow==2.8.0

# Visualization & Interface
streamlit>=1.22.0
matplotlib>=3.3.0
seaborn>=0.11.0
plotly>=5.3.0

# Notebooks & Development
jupyter>=1.0.0
notebook>=6.1.0

# API & Deployment
fastapi>=0.68.0
uvicorn>=0.15.0

# Utils
python-dotenv>=0.19.0
pytest>=6.2.5

# Added from the code block
pyyaml
